feat(Send): adds methods and tests
