+++ src/lib/Send/index.js
@@ -1,0 +1,14 @@
+export default function Send(rpc) {
+  return ({
+    wallet,
+    source,
+    destination,
+    amount,
+  }) =>
+    rpc('send', {
+      wallet,
+      source,
+      destination,
+      amount,
+    });
+}
