--- src/lib/Sending/index.js
@@ -1,6 +1,0 @@
-export default function Sending(rpc) {
-
-  return {
-
-  };
-}
